feat: remove custom action
BREAKING CHANGE: WIP used to allow for custom overrides of a status using a [check run action](https://developer.github.com/changes/2018-05-23-request-actions-on-checks/). But it was barely used and was the sole reason why WIP required write access to pull requests. Because of these reasons I decided to remove the custom action feature